sqlbase: introduce DatabaseDescriptorInterface and ImmutableDatabaseDescriptor
Next we'll stop having the raw descriptor implement that interface.

Release note: None
