+++ pkg/sql/sqlbase/database.go
@@ -9,14 +9,3 @@
 // licenses/APL.txt.
 
 package sqlbase
-
-import "github.com/cockroachdb/cockroach/pkg/sql/sem/tree"
-
-// MakeDatabaseDesc constructs a DatabaseDescriptor from an AST node.
-func MakeDatabaseDesc(id ID, p *tree.CreateDatabase) DatabaseDescriptor {
-	return DatabaseDescriptor{
-		ID:         id,
-		Name:       string(p.Name),
-		Privileges: NewDefaultPrivilegeDescriptor(),
-	}
-}
